IiifController#image_info: specify 'maxWidth' in 'profile' property when image is not downloadable
also, tests:
pull a more generally applicable allow out of the individual tile size test, give it parameter expectations, and put it in a before block in the '#metadata' section, so that all those test cases default to allowing download (previously didn't matter either way for the extant 'profile' test, but matters after this change).  add a test that looks for 'maxWidth' in the 'profile' section when the image isn't downloadable.